requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cohere==5.13.4
fastavro==1.10.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
load-dotenv==0.1.0
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
parameterized==0.9.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
